few fixes and comments added

+++ protocol/v2/ssv/runner/committee.go
@@ -229,7 +229,7 @@ func (cr *CommitteeRunner) ProcessConsensus(logger *zap.Logger, msg *spectypes.S
 			if err != nil {
 				return errors.Wrap(err, "failed to hash attestation data")
 			}
-			logger.Debug("signed attestation data", zap.Int("validator_index", int(duty.ValidatorIndex)),
+			logger.Debug("signed attestation data", zap.Uint64("validator_index", uint64(duty.ValidatorIndex)),
 				zap.String("pub_key", hex.EncodeToString(duty.PubKey[:])),
 				zap.Any("attestation_data", attestationData),
 				zap.String("attestation_data_root", hex.EncodeToString(adr[:])),
